+++ indico/core/notifications.py
@@ -78,6 +78,7 @@ def _log_email(email, event, module, user, meta=None, summary=None):
         'body': email['body'].strip(),
         'state': 'pending',
         'sent_dt': None,
+        'alternatives': email['alternatives'],
         'attachments': sorted(
             a.get_filename('unnamed') if isinstance(a, MIMEBase) else a[0]
             for a in email['attachments']
@@ -128,7 +129,7 @@ def flush_email_queue():
 
 @make_interceptable
 def make_email(to_list=None, cc_list=None, bcc_list=None, *, sender_address=None, reply_address=None, attachments=None,
-               subject=None, body=None, template=None, html=False):
+               subject=None, body=None, template=None, html=False, alternatives=None):
     """Create an email.
 
     The preferred way to specify the email content is using the
@@ -155,6 +156,8 @@ def make_email(to_list=None, cc_list=None, bcc_list=None, *, sender_address=None
     :param template: A template module containing ``get_subject`` and
                      ``get_body`` macros.
     :param html: ``True`` if the email body is HTML
+    :param alternatives: List of ``(content, mimetype)`` tuples of alternative
+                         representations of the message
     """
     from indico.core.emails import get_actual_sender_address
     if template is not None and (subject is not None or body is not None):
@@ -187,4 +190,5 @@ def make_email(to_list=None, cc_list=None, bcc_list=None, *, sender_address=None
         'subject': subject.strip(),
         'body': body.strip(),
         'html': html,
+        'alternatives': alternatives,
     }