+++ CHANGES.rst
@@ -12,6 +12,9 @@ Improvements
 
 - Add a CAPTCHA and rate limiting to the material package endpoint, and an event
   setting to restrict who can generate one (defaults to managers only) (:pr:`6996`)
+- Add support for custom event reminders with freely chosen subject and body, and
+  allow rich-text for the custom message in standard reminders (:pr:`6989`, thanks
+  :user:`tomako, unconventionaldotdev`)
 
 Bugfixes
 ^^^^^^^^